+++ Req/Req/AddFeedback/FeedbackNavigationItems/WriteFeedbackNavigationItem.swift
@@ -26,18 +26,7 @@ struct WriteFeedbackNavigationItem: View {
             Spacer()
                 .frame(width: 16.0)
             
-            Button {
-                switch changeFeedbackBottomView {
-                case .touchPin, .writeFeedback:
-                    cancelWrtieFeedback()
-                default:
-                    cancelAdjustFeedback()
-                }
-            } label: {
-                Text("취소")
-                    .font(.system(size: 18.0, weight: .semibold))
-                    .foregroundColor(.black)
-            }
+            WriteFeedbackNavigationCancelButton(changeFeedbackBottomView: $changeFeedbackBottomView, title: $title, description: $description, pins: $pins, idCount: $idCount)
             
             Spacer()
             
@@ -49,7 +38,6 @@ struct WriteFeedbackNavigationItem: View {
                     } else if changeFeedbackBottomView == .afterAdjustFeedback {
                         completeAdjustFeedback()
                     }
-                    
                 } label: {
                     Text("확인")
                         .font(.system(size: 18.0, weight: .semibold))
@@ -71,22 +59,6 @@ struct WriteFeedbackNavigationItem: View {
 
 extension WriteFeedbackNavigationItem {
     
-    // 피드백 작성 취소버튼
-    func cancelWrtieFeedback() {
-        // 추가버튼 누른 직후 다시 취소할 경우 동작 안하도록 구현
-        if changeFeedbackBottomView == .writeFeedback {
-            // Pin에 부여한 id에 맞게 취소할 시 배열에서 삭제하기 위함
-            let nowId = idCount - 1
-            pins.remove(at: nowId)
-            idCount = nowId
-            
-            title = ""
-            description = ""
-        }
-        
-        changeFeedbackBottomView = .addFeedback
-    }
-    
     // 피드백 작성 확인버튼
     func completeWrtingFeedback() {
         let lastIndex = pins.endIndex - 1
@@ -97,14 +69,6 @@ extension WriteFeedbackNavigationItem {
         description = ""
         
         changeFeedbackBottomView = .addFeedback
-    }
-    
-    // 피드백 수정 취소버튼
-    func cancelAdjustFeedback() {
-        title = ""
-        description = ""
-        
-        changeFeedbackBottomView = .beforeAdjustFeedback
     }
     
     // 피드백 수정 완료버튼
